refactor: web demo
